bumps yacss version, adds block calculations, style helpers

+++ includes/yabrowser/StyleTree.hh
@@ -49,13 +49,13 @@ public:
   yacss::CSSBaseValue decl_lookup (const std::initializer_list<std::string>, const yacss::CSSBaseValue&) const;
 };
 
-inline unsigned to_px (const yacss::CSSBaseValue& value)
+inline float to_px (const yacss::CSSBaseValue& value)
 {
   if (value.type == yacss::ValueType::Length) {
     return value.get<yacss::LengthValue>().val;
   }
 
-  return 0;
+  return 0.0;
 }
 
 bool selector_matches(const yacss::Selector&, const yahtml::Element&);
